+++ src/NCronJob/Registry/JobDefinition.cs
@@ -13,12 +13,15 @@ internal sealed record JobDefinition(
 {
     private int jobExecutionCount;
 
+    private JobExecutionAttributes JobPolicyMetadata { get; } = JobPolicyMetadata ?? new JobExecutionAttributes(Type);
+
     public CancellationToken CancellationToken { get; set; }
 
     public string JobName { get; } = JobName ?? Type.Name;
 
     /// <summary>
-    /// The JobFullName is used as a unique identifier for the job type including anonymous jobs. This helps with concurrency management.
+    /// The JobFullName is used as a unique identifier for the job type including anonymous jobs.
+    /// This helps with concurrency management.
     /// </summary>
     public string JobFullName => JobName == Type.Name
         ? Type.FullName ?? JobName
@@ -26,9 +29,9 @@ internal sealed record JobDefinition(
 
     public int JobExecutionCount => Interlocked.CompareExchange(ref jobExecutionCount, 0, 0);
 
-    public void IncrementJobExecutionCount() => Interlocked.Increment(ref jobExecutionCount);
+    public RetryPolicyAttribute? RetryPolicy => JobPolicyMetadata.RetryPolicy;
 
-    private JobExecutionAttributes JobPolicyMetadata { get; } = JobPolicyMetadata ?? new JobExecutionAttributes(Type);
-    public RetryPolicyAttribute? RetryPolicy => JobPolicyMetadata?.RetryPolicy;
-    public SupportsConcurrencyAttribute? ConcurrencyPolicy => JobPolicyMetadata?.ConcurrencyPolicy;
+    public SupportsConcurrencyAttribute? ConcurrencyPolicy => JobPolicyMetadata.ConcurrencyPolicy;
+
+    public void IncrementJobExecutionCount() => Interlocked.Increment(ref jobExecutionCount);
 }